ready for vue

+++ app/Http/Controllers/PostController.php
@@ -12,11 +12,24 @@ class PostController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+//---------------------------------------------------
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
+
     public function index()
     {
         //
-        //$post = Post::orderBy('title','desc')->get();
-        $post = Post::orderBy('created_at','desc')->paginate(2);
+        $post = Post::orderBy('title','desc')->paginate(10);
+        // $user_id=auth()->user()->id;
+        // $post = Post::where('user_id',$user_id)->paginate(2);
         return view('posts.index')->with('posts',$post);
     }
 
@@ -43,13 +56,29 @@ class PostController extends Controller
         $this->validate($request,[
             'title'=> 'required',
             'body' => 'required',
+            'cover_image'=>'image|nullable|max:1999',
         ]);
+        //Handle file upload
+        if($request->hasFile('cover_image')){
+            //Get_filname Extension
+                $fileNameWithExt= $request->file('cover_image')->getClientOriginalName();
+            //Get just file name
+                $filename=pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just extension
+                $extension=$request->file('cover_image')->getClientOriginalExtension();
+            //File Name to sTore
+            $fileNameToStore=$filename.'_'.time().'.'.$extension;
+            //Upload image
+            $path=$request->file('cover_image')->storeAs('public/cover_images',  $fileNameToStore);
+        }else{
+            $fileNameToStore='noimage.jpg';
+        }
         //Create Post
-
         $post= new POST;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image=$fileNameToStore;
         $post->save();
         return redirect('/posts')->with('success', 'Post Created');
     }
@@ -75,8 +104,13 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        
         //
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id !==$post->user->id){
+            return redirect('/posts')->with('error', 'Unauthorized Action');
+        };
         return view('posts.edit')->with('post', $post);
     }
 
@@ -90,7 +124,19 @@ class PostController extends Controller
     public function update(Request $request, $id)
     {
         //
-
+         //Handle file upload
+         if($request->hasFile('cover_image')){
+            //Get_filname Extension
+                $fileNameWithExt= $request->file('cover_image')->getClientOriginalName();
+            //Get just file name
+                $filename=pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just extension
+                $extension=$request->file('cover_image')->getClientOriginalExtension();
+            //File Name to sTore
+            $fileNameToStore=$filename.'_'.time().'.'.$extension;
+            //Upload image
+            $path=$request->file('cover_image')->storeAs('public/cover_images',  $fileNameToStore);
+        }
         $this->validate($request,[
             'title'=> 'required',
             'body' => 'required',
@@ -99,6 +145,9 @@ class PostController extends Controller
         $post= Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         return redirect('/posts')->with('success', 'Post Updated');
     }
@@ -112,6 +161,9 @@ class PostController extends Controller
     public function destroy($id)
     {
         //
+        if(auth()->user()->$id!==$post->user_id){
+            return redirect('/post')->with('error', 'Unauthorized Page');
+        }
 
         $post = Post::find($id);
         $post->delete();
